// src/payload/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface HomePage {
  id: number;
  heroSlides?:
    | {
        title: string;
        highlightWord?: string | null;
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  scrollCards?:
    | {
        title: string;
        description: string;
        image?: (number | null) | Media;
        buttonText?: string | null;
        id?: string | null;
      }[]
    | null;
  resultsHighlight?:
    | {
        title: string;
        image?: (number | null) | Media;
        yearlyData?:
          | {
              academicYear: string;
              schoolResult: string;
              boardResult: string;
              id?: string | null;
            }[]
          | null;
        id?: string | null;
      }[]
    | null;
  contactUs: {
    schoolAddress: {
      street: string;
      city: string;
    };
    primaryContact: {
      phone: string;
    };
    locationInfo: {
      /**
       * Paste the "src" URL from the Google Maps iframe embed code.
       */
      mapEmbedUrl: string;
    };
    responseTime?: {
      admissionResponse?: string | null;
    };
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}